Move syncModelWithAttributeConfiguration into utility file so it can be used without a GraphController.
